Add info.rb to embedded output
When the embed option (-E) is used, racc adds the contents of
`parser.rb` to the output parser so that the parser can be run by
itself without requiring racc. But `parser.rb` requires `info.rb`,
creating an error when the output parser is run.

These changes add the contents of `info.rb` to the output before the
embedded parser. I tried to stick to the current coding style.