Use hoe-bundler to create Gemfile
This is for Travis CI, so it can install all the dependencies before running
the tests. hoe-travis is supposed to do this, but it doesn't work, as it uses
a Rake task, and Rake doesn't run at all until the required dependencies are
installed.

Bundler users (like me) will also appreciate being able to use `bundle install`
to get all the dependencies needed to work on Racc.

Remove the `isolate` gem, as Bundler does everything it does. Also, the 2 don't
play well together when running tests.

In future, when any dependencies are changed, make sure to run
`rake bundler:gemfile` and commit the results.