Remove unused '_' alias from grammar definition DSL
Racc contains a DSL for defining a grammar in Ruby code. This is used to define the
grammar for Racc's own grammar files. (Otherwise, if the grammar for grammar files was
defined in a grammar file, that would lead to a "chicken-or-egg" circular dependency...)

This DSL is completely unused outside of Racc itself, so it doesn't need extra features
which are not used within Racc.