Remove unused DefinitionEnv#precedence_table method
As with the previous couple of commits, this is part of a DSL which is solely used
for defining the grammar of Racc grammar files. But it is never used in that
grammar definition.

If the "grammar-file grammar" is amended later and needs a precedence table, we
can reintroduce this code. It's still there in the git commit history, anyways...