Greatly simplify Grammar#compute_useless
It's not necessary to compute Rule#useless? here. Actually, which rules are
"useless" is determined later, when building the action table. (It's all the
rules which will never be used for a reduction, because in every instance
where they *could* be used, there is another rule which has higher precedence.)

Useless Syms are those which can never be part of a valid parse tree, because
there is no sequence of rules which lead from the start symbol to that symbol.
We can compute these more simply than what the legacy code did.