Remove undocumented (and never used) user code block file inclusion feature
Racc has (or had) a feature, which probably no-one ever knew about, except the
original author. If you wrote a user code block like this:

    ---- header=path/to/a/file.rb and/another/file.rb

...then the contents of those files would be copied into the user code.

This was not documented. I have looked through the repos of every single gem
on Rubygems which has a dependency on Racc, checked all of their grammars,
and did not find a single example of anyone using this feature.

Therefore, nuke it.