Rename Grammar#init to #finished!
`init` is a misleading name. It implies that the method is called just when
creating a new Grammar object and preparing it for use. Actually, it is called
after all the rules have already been defined, in order to perform some
computations which can only be done when all the rule definitions are available.