Remove useless useless_rules code (can't believe I didn't notice earlier)
When first reading the code through, I thought that States#check_useless in
state.rb set Rule#useless. After another look... I discover that *nothing*
sets Rule#useless.

Racc will never print warnings about 'useless rules', because it doesn't check
for useless rules!

Could a rule be... useless? Sure, it could, if there was no chance that it
would ever be used in a successful parse. Can we detect this? Yeah, I think we
can. BUT, if that is desired, it will be much easier to do it once all the
code is nice and clean and readable.

For now, just remove the useless code.