Remove unused 'fastcache' extension
This (very small) extension was just a hash table for LALR states. It's not
used anywhere. It looks like the author wanted to make Racc run a bit faster
by using something with higher performance than a plain Ruby Hash... but
storing and retrieving States from a Hash is not a bottleneck for performance
anyway.