[Bugfix] Don't infinite-loop if error recovery is attempted at EOF
There is a well-known bug with Racc which often makes it 'hang' when an error
occurs in the input. For example, the bug is discussed here:

https://github.com/whitequark/parser/issues/93

The problem occurs when error recovery is attempted at EOF. To avoid an infinite
loop, error recovery should always 1) consume some of the input by shifting, or
2) consume some of the stack by reducing. Normally, Racc's error recovery does
number 1. In this way, if it gets into a loop of hitting an error, trying to
recover, then hitting another error indefinitely, it should eventually run out
of input and stop.

Except... Racc's code for getting a token from the input has a guard clause
which makes it do nothing when at EOF. Makes sense, but the code for shifting
*still* behaves as if a token had been successfully consumed and decrements
@racc_error_status.

When error recovery is entered again, it aborts the parse if at EOF... but only
if @racc_error_status == 3. In other words: only if no token was successfully
shifted since the last error was hit.

This easily leads to an infinite loop where Racc hits an error at EOF, tries
error recovery, pops a state off the stack and goes back to the previous state,
tries to continue the parse, tries to shift (but gets nothing because of EOF),
the shifting code still looks up the next state in its action table and pushes
it on the stack (restoring the state which the error recovery code discarded),
hits another error, tries to recover, and so on ad infinitum.

To avoid this, do not decrement @racc_error_status on shifting, if the 'shift'
didn't actually shift anything due to being at EOF. I am almost certain that
this is what the original author actually intended.