Add regression test based on cadenza.y from cadenza gem
From the author, William Howard:

    Yes please feel free to include Cadenza's grammar in your test suite.

The original project repo and homepage does not include any copyright notice,
nor does it include a LICENSE file. The homepage simply states that the code is
released under an MIT license. Therefore, add a note to the same effect at the
top of the grammar file, along with links to the author's GitHub page and the
original project repo.