Reimplement GrammarFileScanner
The new GrammarFileScanner has 2 parts: a Ragel-generated scanner for Racc
grammars outside of the action blocks, and a Ripper-based scanner for the Ruby
code in the action blocks.

The new (combined) scanner is more concise and far more readable than the old
one, and faster to boot. Additionally, it can handle all kinds of Ruby
constructs correctly, even the tricky ones which Racc would freak out on before:

- =begin/=end comments
- %r{} regexp literals
- <<HEREDOCS, even more than one on the same line
- // regexp literals and / for division are correctly distinguished, ALWAYS
- nested #{} interpolation, including <<HEREDOCS inside #{} inside...
- ...and all the rest of Ruby's wild and wooly syntax...

The use of Ripper makes it smokin' fast, but also adds a dependency on the
current Ruby interpreter. If you are running Racc on Ruby 2.2, and want to
compile a grammar which uses Ruby 1.8- only syntax in the action blocks, you
will be out of luck. (However, you would probably have been out of luck before
too, simply because the range of constructs recognized by the Racc scanner was
limited.)

Another upshot of using Ripper is that Racc will catch bad syntax in action
blocks, in many cases where it would have passed it through before. (So you
would only have discovered your mistake when running the generated parser.)

Better diagnostics are coming soon.

The new RubyLexer has been tested on a corpus of 1.4 million lines of Ruby code,
and burned through it with no problems.