Remove DebugFlags
Good diagnostic/debugging facilities are important for any program of even
moderate complexity. So why axe the debug flags?

- The usage help itself says "don't use them".
- The information they log is extremely arcane, including state numbers, etc.
- They are undocumented.
- There is almost no test coverage for them.

Before long, we are going to add much better diagnostics and visualizations
of internal parser state.

In the meantime, our test suites still rely on some of the debugging information
logged by -Do. Move that functionality to -v instead, so the tests keep working.
Again, this is only temporary and we will have a better solution later.