Remove --line-convert options
These options cause Racc to emit `module_eval` blocks with 'filename' and
'line number' arguments, so that if an uncaught exception, etc. occurs inside
a Racc action block, or a user code block, Ruby will report the line number in
the original grammar file, rather than in the generated parser file.

This is clearly better for users. So rather than bloating the UI with CLI
options, it's better to simply "do the right thing" all the time. (At least
when it is clear what the "right thing" is, and I think it is clear.)

Previously, converting line numbers was the default in action blocks, and in
some user code blocks, but not all. Now it is done for ALL user code blocks.

This required a couple tweaks to ParserFileGenerator to keep all the tests
passing -- one to make sure that the delimiters used for each block would always
be consistent (for regression tests), and one to ensure that the generated
source could be instance_eval'd inside a module, and classes defined in header
or footer blocks would be available inside the generated parser class.