Drop -P option to 'profile' Racc
The -P, or 'profile', option would make Racc time each stage of the
generation process and print a summary afterward. This is only useful to the
developers of Racc, and not to users, so it doesn't need its own CLI option.
Rather, set an environment variable called PROFILE if you want to see that
summary.