No more Racc '.output' logs
Racc used to output 2 different debug logs: one under ./log, and one which
was named after the input file with .output appended.

The first one has already been removed, and the second one is quickly becoming
redundant. Very soon, we will have DOT visualization of the state graph, and
then it will be completely redundant. So just remove it now.