Show overall parse context where S/R or R/R conflicts occur (verbose mode only)
This is a very powerful tool in figuring out why a S/R or R/R conflict occurs
in a certain part of the grammar.

Previously, Racc would already show a sample of input which would lead to the
S/R or R/R conflict, and list the conflicting rules. This was a BIG help, but
sometimes it is still very tricky to figure out WHY Racc thinks that a certain
reduce rule applies, when it seems "obvious" that only the shift rule should
apply and that following the reduce rule could never lead to a successful parse.

Is this just because the LALR algorithm is too "dumb" to figure out that the
reduce rule is useless in a certain state? Or is there some obscure combination
of inputs going forward, which could actually lead to a successful parse by
applying the reduce rule?

With --verbose, a listing of location pointers (rules and a position in the
rule indicated by a dot) will be printed, showing the overall parse context
which following the "sample input leading to conflict" would result in. From
the start rules going down, each rule which the parser could be "in" after
passing over the sample input is shown.

This can help you see the "paths forward" which the shift and reduce rules
could lead to, if indeed there is a possible "path forward" through each of
them. If not, then that simply means that the LALR algorithm is too weak to
determine exactly which rules could validly be applied at a certain point in
the parse.