Remove y2racc and racc2y
It seems unlikely that there is really any demand for these scripts. They are
in need of code review, and not covered by any tests.

If enough people ask for them later, we may "resurrect" them, do the code
review, and write tests. Otherwise, if there are only a very small number of
people who want to use them, such people can use the 1.4 gem.