Introduce Warnings class
This is a container for categorized warnings, which can be queries as to
whether there are any warnings pertaining to a certain rule, symbol, or state.

This will be used to avoid registering warnings which describe automaton states,
in cases where we have a more specific warning which describes a problem with
a grammar rule.