Rename Sym#dummy? to Sym#generated?
Each grammar has one symbol called the 'dummy' symbol. But, there was also a
`Sym#dummy?` method which would return true for other symbols, too!

Really, the point of this method is whether the symbol was automatically
generated, or whether it appeared explicitly in the input file. So name the
method accordingly.