Embed info.rb and exception.rb in generated parser when -E option used
racc/parser.rb requires these 2 files. If someone generates a parser with
'embedded runtime', and the parser is run on a machine which doesn't have
the racc gem installed, it will fail due to the absence of these files.

-E should embed *everything* which is needed for a standalone parser.

There is already a test case which uses -E: now, I have made sure that
Racc is *not* added to the Ruby load path when running that case.