Fix test for --executable when tests are run using 'bundle exec'
According to the Bundler documentation, "Any Ruby code that opens a subshell
(like system, backticks, or %x{}) will automatically use the current Bundler environment."
Since that is so, it's not necessary to generate executable parser files with
"bundle exec ruby" in the shebang when running tests using 'bundle exec'.

Good thing too, because Linux handles the shebang in a bad way. In Linux, this:

  #!bundle exec ruby

Is the same as running your script as:

  bundle "exec ruby" script.rb

...Which obviously doesn't work. In other words, Linux only allows a single argument
to be included in a shebang line (it doesn't split the arguments on whitespace).