Fix 'escapednl' scanner test with bad Ruby syntax
It turns out that the Ruby syntax used in this test file was always bad, but the test
happened to pass on the version of Ruby which was current at that time.

Ruby contains a library called Ripper, which is based on Ruby's built-in parser and lexer
and enables them to be invoked programmatically. The trouble with Ripper is that as the
Ruby core team changes the parser code, the output from Ripper (even when run on the same
input) may also change. Even seemingly innocuous changes to Ruby's parser may change
Ripper in a way which breaks code which relies on it.

This test file is a good example. On Ruby 2.2.8 and 2.4.2, Racc is able to process it
without error. But on Ruby 2.3.5, the output from Ripper is different and the test fails.

In any case, the Ruby syntax is incorrect. The intention was to ensure that Racc works
when action blocks contain correct Ruby code, not incorrect. So the bad part can just
be removed.