Add GitHub Actions workflow for deploying Racc documentation to GitHub Pages
This PR will allow you to review Racc's documentation in the following GH pages:
https://ruby.github.io/racc/

### Motivations

Currently, the Racc documentation is available below, but does not appear to be maintained. Also, there appears to be no way to check it on the web.
https://github.com/ruby/racc/blob/213e7edd6e60631b2b126008378c8ce61783d4b4/doc

We would like to deploy the documentation to GitHub pages so that people can view the Racc documentation on the web.

### Future work

Currently, we would like to organize and provide documentation on the web for the docs under docs.